Replace OpenAI with Gemini AI for chatbot functionality
