+++ api/index.js
@@ -12,7 +12,7 @@ import contactRouter from "./routes/contact.route.js";
 import wishlistRouter from "./routes/wishlist.route.js";
 import notificationRouter from "./routes/notification.route.js";
 import reviewRouter from "./routes/review.route.js";
-import openaiRouter from "./routes/openai.route.js";
+import geminiRouter from "./routes/gemini.route.js";
 import cors from 'cors';
 import cookieParser from 'cookie-parser';
 import http from 'http';
@@ -234,7 +234,7 @@ app.use("/api/contact", contactRouter);
 app.use("/api/wishlist", wishlistRouter);
 app.use("/api/notifications", notificationRouter);
 app.use("/api/review", reviewRouter);
-app.use("/api/openai", openaiRouter);
+app.use("/api/gemini", geminiRouter);
 
 // Global error handler (should be after all routes)
 app.use((err, req, res, next) => {
