--- api/routes/openai.route.js
@@ -1,8 +1,0 @@
-import express from 'express';
-import { chatWithOpenAI } from '../controllers/openai.controller.js';
-
-const router = express.Router();
-
-router.post('/chat', chatWithOpenAI);
-
-export default router;
