+++ package.json
@@ -34,7 +34,7 @@
     "mongoose": "^8.12.1",
     "multer": "^2.0.1",
     "nodemon": "^3.1.10",
-    "openai": "^5.12.2",
+    "@google/genai": "^1.13.0",
     "react-hot-toast": "^2.5.2",
     "react-swipeable": "^7.0.2",
     "socket.io": "^4.8.1"
